Fixes #53: Fixes 'Comments' view in Offline Mode
While working on the REST fallback for offline mode, I neglected to
refactor item fetching for the newer promises model. This mean’t we
weren’t correctly populating stores that would be used in this mode.
This has been confirmed as now working with this fix. Deploy with fix
running at https://react-hn-comments.appspot.com/